Crate string_cache_shared

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Constants

DYNAMIC_TAG 	 [
Experimental
]

ENTRY_ALIGNMENT 	 [
Experimental
]

INLINE_TAG 	 [
Experimental
]

MAX_INLINE_LEN 	 [
Experimental
]

STATIC_SHIFT_BITS 	 [
Experimental
]

STATIC_TAG 	 [
Experimental
]

TAG_MASK 	 [
Experimental
]

Functions

pack_static 	 [
Experimental
]