Function string_cache_shared::inline_orig_bytes

pub unsafe fn inline_orig_bytes<'a>(data: &'a u64) -> &'a [u8]

For as_slice on inline atoms, we need a pointer into the original string contents.

It's undefined behavior to call this on a non-inline atom!!